requirements.txt
requests==2.26.0
pickle-mixin==1.0.2
GitPython==3.1.24